+++ __tests__/services/pb/world_engine.service.spec.ts
@@ -8,7 +8,7 @@ import {
 } from '../../../proto/world-engine.pb';
 import { CLIENT_ID } from '../../../src/common/constants';
 import { WorldEngineService } from '../../../src/services/pb/world_engine.service';
-import { createAgent, session, user } from '../../helpers';
+import { createAgent, previousState, session, user } from '../../helpers';
 const SCENE = v4();
 
 const agents = [createAgent(), createAgent()];
@@ -220,5 +220,32 @@ describe('load scene', () => {
       fieldId: user.profile.fields[0].id,
       fieldValue: user.profile.fields[0].value,
     });
+  });
+
+  test('should send previous state', async () => {
+    const mockLoadScene = jest.fn(
+      (_req: LoadSceneRequest, _initReq?: fm.InitReq) => {
+        return Promise.resolve({ agents });
+      },
+    );
+
+    WorldEngine.LoadScene = mockLoadScene;
+
+    await client.loadScene({
+      config: {
+        capabilities,
+        connection: {
+          gateway: { hostname: 'examples.com', ssl: true },
+        },
+      },
+      name: SCENE,
+      session,
+      sessionContinuation: { previousState },
+      user,
+    });
+
+    expect(
+      mockLoadScene.mock.calls[0][0].sessionContinuation.previousState,
+    ).toEqual(previousState);
   });
 });