+++ src/entities/continuation/session_continuation.entity.ts
@@ -1,0 +1,11 @@
+export interface SessionContinuationProps {
+  previousState?: Uint8Array;
+}
+
+export class SessionContinuation {
+  readonly previousState: Uint8Array | undefined;
+
+  constructor(props: SessionContinuationProps) {
+    this.previousState = props.previousState;
+  }
+}