+++ src/services/pb/world_engine.service.ts
@@ -8,6 +8,7 @@ import {
   SessionToken,
   User,
 } from '../../common/data_structures';
+import { SessionContinuation } from '../../entities/continuation/session_continuation.entity';
 import { PbService } from './pb.service';
 
 const INWORLD_USER_ID = 'inworldUserId';
@@ -18,12 +19,21 @@ export interface LoadSceneProps {
   user?: User;
   config: InternalClientConfiguration;
   session: SessionToken;
+  sessionContinuation?: SessionContinuation;
   sceneProps?: ExtensionLoadSceneProps;
 }
 
 export class WorldEngineService extends PbService {
   async loadScene(props: LoadSceneProps) {
-    const { client, config, name, sceneProps, session, user } = props;
+    const {
+      client,
+      config,
+      name,
+      sceneProps,
+      session,
+      sessionContinuation,
+      user,
+    } = props;
     const { id, fullName, profile } = user;
 
     return this.request(config, session, WorldEngine.LoadScene, {
@@ -44,6 +54,11 @@ export class WorldEngineService extends PbService {
               ({ id: fieldId, value: fieldValue }) => ({ fieldId, fieldValue }),
             ),
           },
+        },
+      }),
+      ...(sessionContinuation?.previousState && {
+        sessionContinuation: {
+          previousState: sessionContinuation.previousState,
         },
       }),
       ...sceneProps,